+++ CHANGELOG.md
@@ -6,7 +6,11 @@ This is a list of manually mantained changes and updates for each version.
 
 ## Added
 
-None
+#### Debug mode added
+
+`gracenode` can optionally run your application in debug mode.
+
+For more details please read <a href="https://https://github.com/voltrue2/gracenode#debug-mode">here</a>.
 
 ## Changed